Move fn_decl retrieval to utils dir
