+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/utils/fn_decls.rs
@@ -1,8 +1,6 @@
 use quote::format_ident;
-use swc_ecma_ast::{FnDecl, Program};
+use swc_ecma_ast::{ExportDecl, FnDecl, Module, ModuleDecl, Program};
 use syn::Ident;
-
-use crate::generators::canister_methods::get_ast_fn_decls_from_programs;
 
 pub enum CanisterMethodType {
     Heartbeat,
@@ -49,6 +47,51 @@ pub fn get_param_name_idents(fn_decl: &FnDecl) -> Vec<Ident> {
             )
         })
         .collect()
+}
+
+
+pub fn get_ast_fn_decls_from_programs(programs: &Vec<Program>) -> Vec<FnDecl> {
+    programs.iter().fold(vec![], |acc, program| {
+        let ast_fn_decls = get_ast_fn_decls_from_program(program);
+
+        vec![acc, ast_fn_decls].concat()
+    })
+}
+
+fn get_ast_fn_decls_from_program(program: &Program) -> Vec<FnDecl> {
+    match program {
+        Program::Module(module) => {
+            let export_decls = get_export_decls(module);
+
+            let fn_decls: Vec<FnDecl> = export_decls
+                .iter()
+                .filter(|export_decl| export_decl.decl.is_fn_decl())
+                .map(|export_decl| export_decl.decl.as_fn_decl().unwrap().clone())
+                .collect();
+
+            fn_decls
+        }
+        Program::Script(_) => {
+            vec![]
+        }
+    }
+}
+
+fn get_export_decls(module: &Module) -> Vec<ExportDecl> {
+    let module_decls: Vec<ModuleDecl> = module
+        .body
+        .iter()
+        .filter(|module_item| module_item.is_module_decl())
+        .map(|module_item| module_item.as_module_decl().unwrap().clone())
+        .collect();
+
+    let export_decls: Vec<ExportDecl> = module_decls
+        .iter()
+        .filter(|module_decl| module_decl.is_export_decl())
+        .map(|module_decl| module_decl.as_export_decl().unwrap().clone())
+        .collect();
+
+    export_decls
 }
 
 fn is_canister_method_type_fn_decl(
